split compound template helper modules
