+++ .sampo/changesets/issue-373-compound-template-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split compound built-in code template helpers into focused parent, child, and persistence modules while preserving generated scaffold output.
